pyproject.toml
[tool.poetry]
name = "frenetix-motion-planner"
version = "2023.3"
description = "Frenetix Motion Planner: Sampling-based Frenet Planner"
authors = ["Institute of Automotive Technology, Technical University of Munich"]
readme = "README.md"

packages = [
    { include = "frenetix_motion_planner" },
    { include = "cr_scenario_handler" },
    { include = "risk_assessment" },
    { include = "wale_net_lite" },
]

[tool.poetry.dependencies]

# NOTE: commonroad-drivability-checker does not support Python >= 3.11 currently
# (it will be supported in the next release after 2023.1)
# python = ">=3.9,<3.13"

python = ">=3.9,<3.11"

matplotlib = "<=3.7.1"
networkx = "^3.1"
numpy = "^1.24.2"
methodtools = "^0.4.7"
omegaconf = "^2.1.1"
scipy = "^1.7.1"
onnxruntime = "^1.15.1"
pygeos = "^0.14"
imageio = "^2.31.2"
shapely = "^2.0.1"
wheel = ">=0.37.0"
pandas = ">=2.0.1"
triangle = ">=20230923"
"ruamel.yaml" = "^0.17.32"

commonroad-route-planner = ">=2022.3"
commonroad-io = "==2023.2"
commonroad-vehicle-models = "^3.0.2"
commonroad-helper-functions = "^1.0.0"

commonroad-drivability-checker = "2023.1"

frenetix = ">=0.1.0"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
sphinx = ">=2.2.1"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
mypy = "^1.5.1"

[tool.poetry.scripts]
reactive-planner = "main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
